fix(ModuleGameHistoric)
