--- src/app/components/gameHistoric/game-historic.module.ts
@@ -1,10 +1,0 @@
-import { NgModule } from '@angular/core';
-import { CommonModule } from '@angular/common';
-import { GameHistoricRoutingModule } from './game-historic-routing.module';
-import { GameHistoricComponent } from './game-historic/game-historic.component';
-
-@NgModule({
-  declarations: [GameHistoricComponent],
-  imports: [CommonModule, GameHistoricRoutingModule],
-})
-export class GameHistoricModule {}
