--- src/app/components/gameHistoric/game-historic/game-historic.component.html
@@ -1,71 +1,0 @@
-<div class="container">
-  <div class="bg-light m-3 p-5 rounded">
-    <header class="pb-2 mb-2 border-bottom">
-      <span class="fs-4">Histórico de Partidas</span>
-    </header>
-
-    <h1 class="Titulo text-center">Últimas 10 partidas</h1>
-
-    <table class="table">
-      <thead>
-        <tr>
-          <th scope="col">#</th>
-          <th scope="col">Fecha</th>
-          <th scope="col">Intentos</th>
-          <th scope="col">Victoria</th>
-        </tr>
-      </thead>
-      <tbody>
-        <tr *ngFor="let user of lastTenGames; index as i">
-          <th scope="row">{{ i + 1 }}</th>
-          <td
-            [ngStyle]="{
-              color: user.winned === true ? 'green' : 'red'
-            }"
-          >
-            {{ user.date }}
-          </td>
-          <td
-            [ngStyle]="{
-              color: user.winned === true ? 'green' : 'red'
-            }"
-          >
-            {{ user.attempts }}
-          </td>
-          <td
-            [ngStyle]="{
-              color: user.winned === true ? 'green' : 'red'
-            }"
-          >
-            <svg
-              *ngIf="user.winned === true"
-              xmlns="http://www.w3.org/2000/svg"
-              width="16"
-              height="16"
-              fill="currentColor"
-              class="bi bi-check-square-fill"
-              viewBox="0 0 16 16"
-            >
-              <path
-                d="M2 0a2 2 0 0 0-2 2v12a2 2 0 0 0 2 2h12a2 2 0 0 0 2-2V2a2 2 0 0 0-2-2H2zm10.03 4.97a.75.75 0 0 1 .011 1.05l-3.992 4.99a.75.75 0 0 1-1.08.02L4.324 8.384a.75.75 0 1 1 1.06-1.06l2.094 2.093 3.473-4.425a.75.75 0 0 1 1.08-.022z"
-              />
-            </svg>
-            <svg
-              *ngIf="user.winned === false"
-              xmlns="http://www.w3.org/2000/svg"
-              width="20"
-              height="20"
-              fill="currentColor"
-              class="bi bi-file-x-fill"
-              viewBox="0 0 16 16"
-            >
-              <path
-                d="M12 0H4a2 2 0 0 0-2 2v12a2 2 0 0 0 2 2h8a2 2 0 0 0 2-2V2a2 2 0 0 0-2-2zM6.854 6.146 8 7.293l1.146-1.147a.5.5 0 1 1 .708.708L8.707 8l1.147 1.146a.5.5 0 0 1-.708.708L8 8.707 6.854 9.854a.5.5 0 0 1-.708-.708L7.293 8 6.146 6.854a.5.5 0 1 1 .708-.708z"
-              />
-            </svg>
-          </td>
-        </tr>
-      </tbody>
-    </table>
-  </div>
-</div>
