--- src/app/components/gameHistoric/game-historic/game-historic.component.scss
@@ -1,16 +1,0 @@
-.Titulo {
-  font-family: Arial, Helvetica, Sans-serif;
-  color: rgb(6, 6, 6);
-}
-ul {
-  list-style: none;
-}
-.table {
-  border-radius: 23px;
-
-  background-color: rgb(255, 255, 255);
-  color: rgb(17, 17, 17);
-}
-.container {
-  width: 150vb;
-}
