+++ src/app/components/navbar/navbar.component.html
@@ -3,7 +3,6 @@
     class="navbar-brand mx-5"
     id="imgLogo"
     src="../assets/img/icon_wordle.webp"
-    [routerLink]="['/main']"
   />
 
   <div class="navbar-collapse">
