+++ src/app/modules/game-historic-module/components/main/main.component.scss
@@ -1,9 +1,16 @@
 .Titulo {
-    font-family: Arial, Helvetica, Sans-serif;
-    color: rgb(6, 6, 6);
-    text-decoration: underline rgb(14, 14, 14);
-  }
-  ul {
-    list-style: none;
-  }
-  
+  font-family: Arial, Helvetica, Sans-serif;
+  color: rgb(6, 6, 6);
+}
+ul {
+  list-style: none;
+}
+.table {
+  border-radius: 23px;
+
+  background-color: rgb(255, 255, 255);
+  color: rgb(17, 17, 17);
+}
+.container {
+  width: 150vb;
+}
