+++ src/app/modules/game-historic-module/game-historic-module-routing.module.ts
@@ -1,7 +1,12 @@
 import { NgModule } from '@angular/core';
 import { RouterModule, Routes } from '@angular/router';
+import { MainComponent } from './components/main/main.component';
 
-const routes: Routes = [];
+const routes: Routes = [
+  {
+    path: '',component: MainComponent
+  }
+];
 
 @NgModule({
   imports: [RouterModule.forChild(routes)],
